Inserting the structure html text of the site
Laying out the sections and content to be shown in each. Used several placeholders for images and image links, some of which may be replaced by glyphicons. Added the Bootstrap and jQuery CDNs, but have yet to fully implement them. So far I've just used Bootstrap to make the site responsive and format it a bit. The plan right now is to finish html structuring, beautify with Bootstrap/CSS, add more functionality with jQuery/JavaScript, incorporate a MySQL database (visitors log), and add the final text content to the site. For the next step (CSS), I want to look into a proper navbar, scrollspy, affix, and popover. Much more to come...