Progressed functionality, styling, and content
Added dropdowns to the navbar for a couple of the sub-sections and added margins to the titles of these sections to prevent them from being hidden under the navbar after they are scrolled to upon clicking their link in the navbar. Also added a background image and fixed it in place. Inserted a first cut at the website's content and justified the text. Next I will be styling the site more before I iterate on additional or refined functionality. I haven't touched the database aspect yet or the actions that occur upon interacting with some of the elements.