Lowercase ".jpg"
The background image does not appear on the website when loaded from GitHub Pages but does locally. The background image had a capitalized ".JPG" file extension while the CSS code didn't. I doubt this change will fix the problem, but it's a quick test. I have a couple other changes to test if this doesn't work.