Messing with background image and color
I managed to get the background image to appear my ensuring the file extension in the CSS code and the image were both lowercase. I noticed the image is slow to load upon loading the website, which could be for a number of reasons that I will tackle in a later iteration. What I attempted to do this time was see the effects of putting the background image after the background color in the CSS code are on the loading of the website. Currently, the background color is visible while the image slowly loads.