Testing bad background image path
With the background color placed before the background image in the CSS code using the "background-color" property for the color and the "background" property for the background image, the background color does not appear and is overridden by the background image. I now want to test what happens if the background image link is bad. Will the background color then appear?