Responsive background images in mobile browsers
The background images are responsive in my desktop browsers, but not in Safari on my mobile phone. I read that mobile browsers may not support the CSS3 "background-size" property or the "fixed" value for the "background-attachment" property. This commit will test moving the "width" and "height" properties to before the "background" property of one of the images, changing the "max-width" property to "width," changing the "background-attachment" property of one of the background images to "scroll" instead of "fixed," and changing the "background-size" property of another background image to "contain" instead of "cover."