Background images scroll with small screen
Implemented the workaround to make the background images scroll with the page when the browser window is small. This was done to make the background images appear when the website is viewed on a mobile device. Recall that fixed background images are not supported by mobile browsers, generally speaking. This commit will allow me to see if the workaround was successful on my mobile phone.