Added files via upload
The background image workaround was successful in my mobile browser. I implemented the workaround for all background images, aligned the images, formatted the Contacts section, and made the Contacts section in the navbar a dropdown menu. I ran into an issue with links in the navbar. It seems my smooth scroll JavaScript code prevents links in the navbar from working. I will have to revisit this. I removed the smooth scroll feature for now.