Added files via upload
I fixed the smooth scroll feature to worth with the navbar links. I was disabling the default click behavior of the external links. I also fixed the pixel widths for background images to work on larger mobile devices (e.g. ipads). I also changed the position of the background images when they are viewed in a mobile browser since more of the images will be cut off. I also want to see if the slide-in animation works in Firefox. I've been developing and testing using Chrome, where the animation works, but it only works in Firefox when the website is viewed locally as opposed to being pulled from the GitHub Pages web server.