Smooth scroll and slide-in animation
Smooth scroll wasn't working in Firefox, so I updated it and smooth scroll worked. However, the slide-in animation still doesn't work in Firefox. The images fade in as they are supposed to, but don't slide in. The first attempt will be to add "-moz-keyframes" code to the CSS file (I was relying on "-webkit-keyframes" working).