Almost at first static release
I changed the font size of the navbar, finished putting together the images to complement my Technical Thesis section, made the navbar dropdowns drop on hover, maintained the same navbar color when a dropdown tab is clicked, styled the popovers, added the larger-sized images to the Skills section, added links in the About Me section, and iterated on the content. Almost at the first release of the static site. I want to format and comment the code a little first and get feedback from others on the site.