Changes for varying viewport sizes
After visiting the website on a range of viewport sizes, I made a few changes to the styling (and will likely have to make more to get it right). I lessened the padding at the bottom of the title background because the text was not centered vertically on the larger screen. I didn't need as much padding as I had at the bottom. I compressed the images to make the site load faster. I made a first attempt at re-positioning the images to look good for mobile and large viewports. I emphasized "Hover" in the Skills section. I left-aligned the titles that were justified since they looked weird in the mobile browser's small viewport.